Revert "Move Post business logic to mediatr handlers "
This reverts commit 6ea7efb668cc89b2fedf115e5a8ab6cda2b1b589.
